+++ application/views/layout/footer_footer.php
@@ -33,6 +33,8 @@
 
 </script>
 
+<?php $this->load->view('js/lihocos-js'); ?>
+
 </body>
 
 </html>