+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.1.2 (2014-03-24)
+* Better handling of IP cameras (stream/image proxy)
+* Changed home layout to panels instead of tabs
+* Added update instructions (UPDATE.md)
+* General fixes and code improvements
+
 ### 0.1.1 (2014-02-17)
 * Some fixes regarding installation process
 * Added Favicon